Refactor queries and mutations to tests
Co-locate graphql queries with their tests and export them so they
can be used as dependencies by other tests.
