+++ packages/truffle-db/src/workspace/test/contract.spec.ts
@@ -1,8 +1,9 @@
+import gql from "graphql-tag";
 import { generateId, Migrations, WorkspaceClient } from './utils';
-import { Query, Mutation } from './queries';
 
-const { AddSource, AddBytecode, AddCompilation, AddContracts } = Mutation;
-const { GetContract } = Query;
+import { AddSource } from './source.spec';
+import { AddBytecode } from './bytecode.spec';
+import { AddCompilation } from './compilation.spec';
 
 describe("Contract", () => {
   let wsClient;
@@ -93,4 +94,68 @@ describe("Contract", () => {
   });
 
 });
+
+export const GetContract = gql`
+  query getContract($id:ID!){
+      contract(id:$id) {
+        name
+        abi {
+          json
+        }
+        sourceContract {
+          source {
+            contents
+          }
+          ast {
+            json
+          }
+        }
+      }
+  }
+`;
+
+
+export const AddContracts = gql`
+  mutation addContracts($contractName: String, $compilationId: ID!, $bytecodeId:ID!, $abi:String!) {
+    contractsAdd(input: {
+      contracts: [{
+        name: $contractName
+        abi: {
+          json: $abi
+        }
+        compilation: {
+          id: $compilationId
+        }
+        sourceContract: {
+          index: 0
+        }
+        constructor: {
+          createBytecode: {
+            id: $bytecodeId
+          }
+        }
+      }]
+    }) {
+      contracts {
+        id
+        name
+        sourceContract {
+          name
+          source {
+            contents
+          }
+          ast {
+            json
+          }
+        }
+        constructor {
+          createBytecode {
+            bytes
+          }
+        }
+      }
+    }
+  }
+`;
+
