+++ packages/truffle-db/src/workspace/test/source.spec.ts
@@ -1,9 +1,5 @@
+import gql from "graphql-tag";
 import { generateId, Migrations, WorkspaceClient } from './utils';
-import { Query, Mutation } from './queries';
-
-const { GetSource } = Query;
-const { AddSource } = Mutation;
-
 
 describe("Source", () => {
   let wsClient, addSourceResult
@@ -57,4 +53,31 @@ describe("Source", () => {
   });
 
 });
+
+export const GetSource = gql`
+  query GetSource($id: ID!) {
+    source(id: $id) {
+      id
+      contents
+      sourcePath
+    }
+  }
+`
+
+export const AddSource = gql`
+  mutation AddSource($contents: String!, $sourcePath: String) {
+    sourcesAdd(input: {
+      sources: [
+         {
+           contents: $contents,
+           sourcePath: $sourcePath,
+         }
+      ]
+    }) {
+      sources {
+        id
+      }
+    }
+  }
+`
